==> pages/chapters/14-Bayesian-models-of-perception.qmd <==
---
title: "Bayesian Models of Perception"
description: | 
    Bayesian modelling of perceptual tasks
date: "2022-05-31"
author:
  - name: Andrew Ellis
    url: https://github.com/awellis
    affiliation: Kognitive Psychologie, Wahrnehmung und Methodenlehre, Universität Bern 
    affiliation-url: https://www.kog.psy.unibe.ch
    orcid: 0000-0002-2788-936X
license: CC BY
citation: true
bibliography: ../../bibliography.bib
format:
    html:
        toc: true
        code-link: true
        code-fold: false
        code-tools: true
---

# Bayesian models of perception

Bayesian inference provides a recipe for optimally combining prior knowledge with new sensory observations and it has the potential to provide significant insight into perception (e.g., Kersten, Mamassian, & Yuille, 2004; Knill & Richards, 1996).

@vincentTutorialBayesianModels2015a

- the ‘Bayesian Brain’ hypothesis

In the perceptual domain, the core assertions are that: (a) an observer has an internal mental model which represents the processes that gave rise to their sensory observations, (b) that observers conduct Bayesian inference using their mental model in order to infer probable states of the world from sensory data, and (c) that observers have prior beliefs over states of the world.

## Bayesian observer models

If we accept that we do not have direct access to the true state of the world, and the data are insufficient to provide this knowledge unambiguously, then Bayesian inference becomes appealing as the only principled solution to the inverse problem.

### Optimal vs. suboptimal Bayesian observers

The Bayesian optimal observer approach would claim that observer’s mental models of the world are veridical, and that their prior beliefs are matched to the statistics of the environment. In hypothetical ideal observer models, observers are given precise knowledge of certain aspects of the world in order to calculate the theoretical maximum performance.

The first reason for this suboptimality could be that an observer’s internal mental model of the world is mismatched with the actual generative process (the real world) which gave rise to their sensory observations.


## References
